move AutoCompleteItemComponent to AutoComplete::Item
