+++ app/components/primer/auto_complete.rb
@@ -40,10 +40,10 @@ module Primer
     #   <%= render(Primer::AutoComplete.new(src: "/users/search", id: "user-popup", position: :relative)) do |c| %>
     #     <% c.input(type: :text, name: "input") %>
     #     <% c.results do %>
-    #       <%= render(Primer::AutoCompleteItemComponent.new(selected: true, value: "value")) do |c| %>
+    #       <%= render(Primer::AutoComplete::Item.new(selected: true, value: "value")) do |c| %>
     #         Selected
     #       <% end %>
-    #       <%= render(Primer::AutoCompleteItemComponent.new(value: "value")) do |c| %>
+    #       <%= render(Primer::AutoComplete::Item.new(value: "value")) do |c| %>
     #         Not selected
     #       <% end %>
     #     <% end %>
@@ -53,10 +53,10 @@ module Primer
     #   <%= render(Primer::AutoComplete.new(src: "/users/search", id: "user-popup", position: :relative)) do |c| %>
     #     <% c.input(type: :text, name: "input") %>
     #     <% c.results(classes: "my-custom-class") do %>
-    #       <%= render(Primer::AutoCompleteItemComponent.new(selected: true, value: "value")) do |c| %>
+    #       <%= render(Primer::AutoComplete::Item.new(selected: true, value: "value")) do |c| %>
     #         Selected
     #       <% end %>
-    #       <%= render(Primer::AutoCompleteItemComponent.new(value: "value")) do |c| %>
+    #       <%= render(Primer::AutoComplete::Item.new(value: "value")) do |c| %>
     #         Not selected
     #       <% end %>
     #     <% end %>
@@ -67,10 +67,10 @@ module Primer
     #     <% c.input(type: :text, name: "input") %>
     #     <% c.icon(icon: :search) %>
     #     <% c.results do %>
-    #       <%= render(Primer::AutoCompleteItemComponent.new(selected: true, value: "value")) do |c| %>
+    #       <%= render(Primer::AutoComplete::Item.new(selected: true, value: "value")) do |c| %>
     #         Selected
     #       <% end %>
-    #       <%= render(Primer::AutoCompleteItemComponent.new(value: "value")) do |c| %>
+    #       <%= render(Primer::AutoComplete::Item.new(value: "value")) do |c| %>
     #         Not selected
     #       <% end %>
     #     <% end %>
