+++ demo/app/views/auto_complete_test/index.html.erb
@@ -1,3 +1,3 @@
 <% @fruit_list.each do |fruit| %>
-  <%= render(Primer::AutoCompleteItemComponent.new(value: fruit)) { fruit } %>
+  <%= render(Primer::AutoComplete::Item.new(value: fruit)) { fruit } %>
 <% end %>
