+++ docs/content/components/autocompleteitem.md
@@ -18,10 +18,10 @@ Use AutoCompleteItem to list results of an auto-completed search.
 <Example src="<li role='option' data-autocomplete-value='value' aria-selected='true' class='autocomplete-item '>  Selected</li><li role='option' data-autocomplete-value='value' class='autocomplete-item '>  Not selected</li>" />
 
 ```erb
-<%= render(Primer::AutoCompleteItemComponent.new(selected: true, value: "value")) do |c| %>
+<%= render(Primer::AutoComplete::Item.new(selected: true, value: "value")) do |c| %>
   Selected
 <% end %>
-<%= render(Primer::AutoCompleteItemComponent.new(value: "value")) do |c| %>
+<%= render(Primer::AutoComplete::Item.new(value: "value")) do |c| %>
   Not selected
 <% end %>
 ```
