+++ test/components/auto_complete_test.rb
